+++ lib/rules/prefer-linked-key-with-paren.ts
@@ -3,20 +3,26 @@
  */
 import type { AST as JSONAST } from 'jsonc-eslint-parser'
 import type { AST as YAMLAST } from 'yaml-eslint-parser'
-import { extname } from 'path'
-import { defineCustomBlocksVisitor, getLocaleMessages } from '../utils/index'
-import debugBuilder from 'debug'
 import type { RuleContext, RuleListener } from '../types'
+import type { GetReportOffset } from '../utils/rule'
+import type { CustomBlockVisitorFactory } from '../types/vue-parser-services'
+import { extname } from 'node:path'
+import debugBuilder from 'debug'
+import {
+  createRule,
+  defineCreateVisitorForJson,
+  defineCreateVisitorForYaml
+} from '../utils/rule'
+import { defineCustomBlocksVisitor, getLocaleMessages } from '../utils/index'
 import {
   getMessageSyntaxVersions,
-  getReportIndex,
   NodeTypes
 } from '../utils/message-compiler/utils'
 import { parse } from '../utils/message-compiler/parser'
 import { parse as parseForV8 } from '../utils/message-compiler/parser-v8'
 import { traverseNode } from '../utils/message-compiler/traverser'
-import { createRule } from '../utils/rule'
 import { getFilename, getSourceCode } from '../utils/compat'
+
 const debug = debugBuilder(
   'eslint-plugin-vue-i18n:prefer-linked-key-with-paren'
 )
@@ -30,8 +36,6 @@ function getSingleQuote(node: JSONAST.JSONStringLiteral | YAMLAST.YAMLScalar) {
   }
   return "'"
 }
-
-type GetReportOffset = (offset: number) => number | null
 
 function create(context: RuleContext): RuleListener {
   const filename = getFilename(context)
@@ -141,80 +145,9 @@ function create(context: RuleContext): RuleListener {
       verifyForV8(message, reportNode, getReportOffset)
     }
   }
-  /**
-   * Create node visitor for JSON
-   */
-  function createVisitorForJson(): RuleListener {
-    function verifyExpression(node: JSONAST.JSONExpression) {
-      if (node.type !== 'JSONLiteral' || typeof node.value !== 'string') {
-        return
-      }
-      verifyMessage(node.value, node as JSONAST.JSONStringLiteral, offset =>
-        getReportIndex(node, offset)
-      )
-    }
-    return {
-      JSONProperty(node: JSONAST.JSONProperty) {
-        verifyExpression(node.value)
-      },
-      JSONArrayExpression(node: JSONAST.JSONArrayExpression) {
-        for (const element of node.elements) {
-          if (element) verifyExpression(element)
-        }
-      }
-    }
-  }
 
-  /**
-   * Create node visitor for YAML
-   */
-  function createVisitorForYaml(): RuleListener {
-    const yamlKeyNodes = new Set<YAMLAST.YAMLContent | YAMLAST.YAMLWithMeta>()
-    function withinKey(node: YAMLAST.YAMLNode) {
-      for (const keyNode of yamlKeyNodes) {
-        if (
-          keyNode.range[0] <= node.range[0] &&
-          node.range[0] < keyNode.range[1]
-        ) {
-          return true
-        }
-      }
-      return false
-    }
-    function verifyContent(node: YAMLAST.YAMLContent | YAMLAST.YAMLWithMeta) {
-      const valueNode = node.type === 'YAMLWithMeta' ? node.value : node
-      if (
-        !valueNode ||
-        valueNode.type !== 'YAMLScalar' ||
-        typeof valueNode.value !== 'string'
-      ) {
-        return
-      }
-      verifyMessage(valueNode.value, valueNode, offset =>
-        getReportIndex(valueNode, offset)
-      )
-    }
-    return {
-      YAMLPair(node: YAMLAST.YAMLPair) {
-        if (withinKey(node)) {
-          return
-        }
-        if (node.key != null) {
-          yamlKeyNodes.add(node.key)
-        }
-
-        if (node.value) verifyContent(node.value)
-      },
-      YAMLSequence(node: YAMLAST.YAMLSequence) {
-        if (withinKey(node)) {
-          return
-        }
-        for (const entry of node.entries) {
-          if (entry) verifyContent(entry)
-        }
-      }
-    }
-  }
+  const createVisitorForJson = defineCreateVisitorForJson(verifyMessage)
+  const createVisitorForYaml = defineCreateVisitorForYaml(verifyMessage)
 
   if (extname(filename) === '.vue') {
     return defineCustomBlocksVisitor(
@@ -234,9 +167,13 @@ function create(context: RuleContext): RuleListener {
     }
 
     if (sourceCode.parserServices.isJSON) {
-      return createVisitorForJson()
+      return createVisitorForJson(
+        context as Parameters<CustomBlockVisitorFactory>[0]
+      )
     } else if (sourceCode.parserServices.isYAML) {
-      return createVisitorForYaml()
+      return createVisitorForYaml(
+        context as Parameters<CustomBlockVisitorFactory>[0]
+      )
     }
     return {}
   } else {
